Move ComputeReplicaLogging to mz-catalog-types
Move the ComputeReplicaLogging type from mz-compute-types to mz-catalog-types.
The original crate now re-exports the type for backwards compatibility.
