+++ src/compute-types/Cargo.toml
@@ -14,6 +14,7 @@ columnar = "0.11.1"
 columnation = "0.1.0"
 differential-dataflow = "0.18.0"
 itertools = "0.14.0"
+mz-catalog-types = { path = "../catalog-types" }
 mz-dyncfg = { path = "../dyncfg" }
 mz-expr = { path = "../expr" }
 mz-ore = { path = "../ore", features = ["tracing", "metrics"] }
